requirements.txt
aes==1.0.0
asgiref==3.3.1
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
crypto==1.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-import-export==2.5.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
djongo==1.3.6
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
Naked==0.1.31
odfpy==1.4.1
openpyxl==3.0.7
passlib==1.7.4
Pillow==8.2.0
pycodestyle==2.7.0
pycryptodome==3.10.1
pycryptodomex==3.10.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.3
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
